requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
bcrypt==4.1.1

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Date/Time
python-dateutil==2.8.2

# CORS
fastapi-cors==0.0.6

# Excel/PDF Generation
openpyxl==3.1.2
reportlab==4.0.7
pandas==2.1.3

# Background Tasks
apscheduler==3.10.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

# Code Quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Utilities
pytz==2023.3
